nbrain-2025/backend/requirements.txt
fastapi
uvicorn==0.30.1
python-dotenv
pinecone-client
langchain-google-genai
langchain-pinecone==0.1.1
langchain
python-multipart
langchain-text-splitters==0.2.2
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
uvloop==0.19.0
lxml
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
email-validator
pandas
websockets
numpy
httpx
pydantic
deepgram-sdk
elevenlabs==0.2.26
twilio==8.10.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0 
openai>=1.0.0

# Caching
redis==5.0.0
Pillow==10.0.0
dateparser==1.2.0 

# RAG-specific dependencies
flask-sqlalchemy==3.1.1  # For compatibility with Generic RAG models
flask-migrate==4.0.5  # For database migrations compatibility 